Add files via upload
This is an Exploratory Data Analysis project in Coding Ninja's Data Science course.

This dataset has funding information of the Indian startups from January 2015 to August 2017.

Feature Details :

SNo - Serial number Date - Date of funding in format DD/MM/YYYY StartupName - Name of the startup which got funded IndustryVertical - Industry to which the startup belongs SubVertical - Sub-category of the industry type CityLocation - City which the startup is based out of InvestorsName - Name of the investors involved in the funding round InvestmentType - Either Private Equity or Seed Funding AmountInUSD - Funding Amount in USD Remarks - Other information, if any Insights :

Find out what type of startups are getting funded in the last few years? Who are the important investors? What are the hot fields that get a lot of funding these days?